requirements.txt
Pillow==9.5.0
plotly==5.16.0
streamlit==1.25.0
pandas==2.0.3
numpy==1.25.2
folium==0.14.0
matplotlib==3.7.2
matplotlib-inline==0.1.6